Rename module astconv to hir_ty_lowering
Split from the main renaming commit to make git generate a proper diff for ease of reviewing.
